re: change id-based routing to title-based
change all id-based routes to title/name based routes, adjust ll the links leading to those routes, adjust data/portfolio.js to meet consistency throughout the project
