+++ app/(team)/team-details/[name]/page.jsx
@@ -1,28 +1,29 @@
 import MarqueeComponent from "@/components/common/Marquee";
 import Footer8 from "@/components/footers/Footer8";
 import Header3 from "@/components/headers/Header3";
+import Contact2 from "@/components/team/Contact2";
 
-import ServiceDetails from "@/components/service/ServiceDetails";
-import { allFeatures } from "@/data/features";
+import TeamDetails from "@/components/team/TeamDetails";
+import { teamData } from "@/data/team";
 
 //For Static Side Genaration(SSG)
 
 // export async function generateStaticParams() {
-//   return allFeatures.map((elm) => ({
+//   return teamData.map((elm) => ({
 //     id: `${elm.id}`,
 //   }));
 // }
 
 export const metadata = {
-  title:
-    "Service Details || Tecvity - Delivering Creative Technological Solutions",
+  title: "Team Details || Frisk - Creative Agency & Portfolio Nextjs Template",
 };
 
-export default function ServicePageDetails({ params }) {
+export default function TeamPageDetails({ params }) {
   return (
     <>
       <Header3 />
-      <ServiceDetails serviceId={params.id} />
+      <TeamDetails teamName={params.name} />
+      <Contact2 />
       <MarqueeComponent />
       <Footer8 />
     </>
