+++ components/service/ServiceDetails.jsx
@@ -4,9 +4,10 @@ import { allFeatures } from "@/data/features";
 import Image from "next/image";
 import Link from "next/link";
 
-export default function ServiceDetails({ serviceId }) {
+export default function ServiceDetails({ serviceTitle }) {
+  const decodedServiceTitle = decodeURIComponent(serviceTitle);
   const serviceItem =
-    allFeatures.filter((elm) => elm.id == serviceId)[0] || allFeatures[1];
+    allFeatures.filter((elm) => elm.title == decodedServiceTitle)[0] || allFeatures[1];
 
     const getRandomIcon = () => {
       const icons = [
